修改了一些由merge引发的错误

+++ bbs-springboot/bbs-article/bbs-article-facade/src/main/java/com/liang/bbs/article/facade/server/CommentService.java
@@ -8,9 +8,6 @@ import com.liang.nansheng.common.auth.UserSsoDTO;
 import java.time.LocalDateTime;
 import java.util.List;
 
-/**
- * @date 2022/4/6 14:33
- */
 public interface CommentService {
     /**
      * 获取文章的评论信息
@@ -76,7 +73,7 @@ public interface CommentService {
      * @param commentId
      * @return
      */
-    Boolean delete(Integer commentId);
+    Boolean delete(Long commentId);
 
     /**
      * 通过父级ID获取子级评论信息
@@ -85,7 +82,7 @@ public interface CommentService {
      * @param preId
      * @return
      */
-    void getAllChildrenByPreId(List<CommentDTO> result, Integer preId);
+    void getAllChildrenByPreId(List<CommentDTO> result, Long preId);
 
     /**
      * 获取评论id获取文章id
@@ -93,7 +90,7 @@ public interface CommentService {
      * @param commentId
      * @return
      */
-    Integer getArticleIdByCommentId(Integer commentId);
+    Integer getArticleIdByCommentId(Long commentId);
 
     /**
      * 通过批量id获取评论信息
@@ -101,5 +98,5 @@ public interface CommentService {
      * @param commentId
      * @return
      */
-    CommentDTO getById(Integer commentId);
+    CommentDTO getById(Long commentId);
 }
